# public/openapi.yaml
openapi: 3.0.3
info:
  title: BakedBot Partner API
  description: |
    The BakedBot Partner API lets cannabis POS systems and e-commerce platforms connect dispensary
    transaction data to BakedBot's retention and loyalty engine.

    Integrating takes four steps:
    1. Request API access at **bakedbot.ai/developers**
    2. Receive your org-scoped API key
    3. Configure your POS to POST completed transactions to `/api/v1/integrations/pos/transactions/completed`
    4. Optionally enroll customers and look up loyalty profiles at checkout

    All endpoints require a Bearer token issued by BakedBot. Keys are scoped to permissions —
    the required scope is noted per endpoint.

    **Base URL:** `https://bakedbot.ai`
  version: "1.0.0"
  contact:
    name: BakedBot Integration Support
    email: martez@bakedbot.ai
    url: https://bakedbot.ai/developers
  license:
    name: Proprietary
servers:
  - url: https://bakedbot.ai
    description: Production
security:
  - BearerAuth: []
...
paths:
  /api/v1/integrations/pos/transactions/completed:
    post:
      tags: [Transactions]
      summary: Record a completed POS transaction
      description: |
        POST this endpoint when a Dutchie transaction is finalized. BakedBot will:
        - Award loyalty points to the matched member (1 point per dollar spent)
        - Mark the active visit session as completed
        - Fire behavior-triggered retention campaigns (win-back, loyalty milestone, product replenishment)

        BakedBot matches the transaction to an open loyalty session first by `posCartRef`, then
        falls back to `customerPhone`. If no session matches, the transaction is recorded but
        loyalty actions are skipped.
      operationId: recordCompletedTransaction
      security:
        - BearerAuth: [write:transactions]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransactionCompletedRequest'
            example:
              organizationId: "org_thrive_syracuse"
              storeId: "store_main"
              posCartRef: "cart_abc123"
              posTransactionRef: "txn_dutchie_98765"
              customerPhone: "+13155550100"
              totalCents: 8500
              subtotalCents: 8500
              discountCents: 0
              items:
                - skuId: "sku_blue_dream_1g"
                  name: "Blue Dream 1g"
                  quantity: 2
                  totalCents: 4000
      responses:
        '200':
          description: Transaction recorded successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionCompletedResponse'
              examples:
                matched:
                  summary: Session matched — points awarded
                  value:
                    success: true
                    sessionId: "vsess_abc123"
                    pointsAwarded: 85
                unmatched:
                  summary: No loyalty session found
                  value:
                    success: true
                    message: "Transaction processed (no loyalty session matched)"
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: |
        API key issued by BakedBot per organization. Pass as `Authorization: Bearer <key>`.
        Keys are scoped to one or more permissions. Request access at bakedbot.ai/developers.
  schemas:
    TransactionCompletedRequest:
      type: object
      required: [organizationId, posTransactionRef, totalCents, subtotalCents]
      properties:
        organizationId:
          type: string
          description: BakedBot org identifier provided at onboarding
        storeId:
          type: string
          description: Location identifier for multi-location dispensaries
        posCartRef:
          type: string
          description: POS cart or order ID — used to match an open loyalty session
        posTransactionRef:
          type: string
          description: Unique POS transaction ID
        customerPhone:
          type: string
          description: Customer phone in E.164 format — fallback session match if posCartRef not found
        customerId:
          type: string
          description: POS-native customer ID for future cross-reference
        totalCents:
          type: integer
          minimum: 0
          description: Transaction total in cents (drives loyalty point calculation)
        subtotalCents:
          type: integer
          minimum: 0
          description: Pre-discount subtotal in cents
        discountCents:
          type: integer
          default: 0
          description: Discount applied in cents
        items:
          type: array
          description: Line items — used for product preference segmentation
          items:
            type: object
            required: [skuId, name, quantity, totalCents]
            properties:
              skuId:
                type: string
              name:
                type: string
              quantity:
                type: integer
              totalCents:
                type: integer
    TransactionCompletedResponse:
      type: object
      properties:
        success:
          type: boolean
        sessionId:
          type: string
          description: Matched loyalty session ID (null if no session found)
        pointsAwarded:
          type: integer
          description: Loyalty points awarded (floor of totalCents / 100)
        message:
          type: string
          description: Present when no loyalty session was matched
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
  responses:
    ValidationError:
      description: Missing or invalid request fields
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error: "posTransactionRef is required"
    Unauthorized:
      description: Missing or invalid API key
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error: "Authorization header with Bearer token required"
    Forbidden:
      description: API key lacks required permission or belongs to a different organization
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error: "Unauthorized: API key does not belong to this organization"